+++ h2gis-functions/src/main/java/org/h2gis/functions/io/dbf/DBFDriverFunction.java
@@ -227,7 +227,6 @@ public class DBFDriverFunction implements DriverFunction {
             if (deleteTables) {
                 Statement stmt = connection.createStatement();
                 stmt.execute("DROP TABLE IF EXISTS " + outputTable);
-
                 stmt.close();
             }
             DBFDriver dbfDriver = new DBFDriver();
@@ -247,6 +246,7 @@ public class DBFDriverFunction implements DriverFunction {
                                 types));
                     }
                     try {
+                        connection.setAutoCommit(false);
                         int columnCount = dbfDriver.getFieldCount();
                         try (PreparedStatement preparedStatement = connection.prepareStatement(
                                 String.format("INSERT INTO %s VALUES ( %s )", outputTable,
@@ -262,6 +262,7 @@ public class DBFDriverFunction implements DriverFunction {
                                 batchSize++;
                                 if (batchSize >= BATCH_MAX_SIZE) {
                                     preparedStatement.executeBatch();
+                                    connection.commit();
                                     preparedStatement.clearBatch();
                                     batchSize = 0;
                                     copyProgress.endStep();
@@ -269,16 +270,20 @@ public class DBFDriverFunction implements DriverFunction {
                             }
                             if (batchSize > 0) {
                                 preparedStatement.executeBatch();
+                                connection.commit();
                                 preparedStatement.clearBatch();
                             }
                         }
                     } catch (Exception ex) {
+                        connection.setAutoCommit(true);
                         connection.createStatement().execute("DROP TABLE IF EXISTS " + outputTable);
                         throw new SQLException(ex.getLocalizedMessage(), ex);
                     }
                 } finally {
+                    connection.setAutoCommit(true);
                     dbfDriver.close();
                     copyProgress.endOfProgress();
+                    connection.setAutoCommit(true);
                 }
                 return new String[]{outputTable};
             }